Make sorted_index an INT

+++ aquadoggo/migrations/20230523135826_alter-operations.sql
@@ -1,3 +1,3 @@
 -- SPDX-License-Identifier: AGPL-3.0-or-later
 
-ALTER TABLE operations_v1 ADD COLUMN sorted_index NUMERIC;
+ALTER TABLE operations_v1 ADD COLUMN sorted_index INT;
